requirements.txt
flask==2.3.3
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0
beautifulsoup4==4.12.3
lxml==4.9.3